+++ lucadeal/lib/luca_deal.rb
@@ -8,6 +8,7 @@ module LucaDeal
   autoload :Contract, 'luca_deal/contract'
   autoload :Fee, 'luca_deal/fee'
   autoload :Invoice, 'luca_deal/invoice'
+  autoload :NoInvoice, 'luca_deal/no_invoice'
   autoload :Product, 'luca_deal/product'
   autoload :Setup, 'luca_deal/setup'
 end